+++ p2panda/tests/spaces.rs
@@ -1,6 +1,15 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+use p2panda::{NetworkId, Node};
 use serde::{Deserialize, Serialize};
+
+async fn spawn_node(network_id: NetworkId) -> Node {
+    p2panda::builder()
+        .network_id(network_id)
+        .spawn()
+        .await
+        .unwrap()
+}
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 struct SecretData {
@@ -19,15 +28,15 @@ mod spaces_api {
     use p2panda_spaces::SpaceEvent;
     use tokio_stream::StreamExt;
 
-    use super::SecretData;
+    use super::{SecretData, spawn_node};
 
     #[tokio::test]
     async fn create_space_for_multiple_members() -> Result<(), Box<dyn std::error::Error>> {
         setup_logging();
 
-        use p2panda::Topic;
+        let network_id = Topic::random().into();
 
-        let panda = p2panda::spawn().await?;
+        let panda = spawn_node(network_id).await;
         let mut panda_system_rx = panda.event_stream().await?;
 
         // Spaces behave like topic-streams, just that they're encrypted towards members.
@@ -48,8 +57,8 @@ mod spaces_api {
         }
 
         // We can manage (nested) groups (useful for multi-device, etc.)
-        let penguin_laptop = p2panda::spawn().await?;
-        let penguin_mobile = p2panda::spawn().await?;
+        let penguin_laptop = spawn_node(network_id).await;
+        let penguin_mobile = spawn_node(network_id).await;
         let mut penguin_mobile_system_rx = penguin_mobile.event_stream().await?;
 
         // Penguin subscribes to the space in order to publish some key bundles.
@@ -280,10 +289,11 @@ mod spaces_api {
     async fn spaces_sync() -> Result<(), Box<dyn std::error::Error>> {
         setup_logging();
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await?;
-        let penguin = p2panda::spawn().await?;
+        let panda = spawn_node(network_id).await;
+        let penguin = spawn_node(network_id).await;
 
         // Penguin subscribes to the space (and publishes a key bundle).
         let (_penguin_space, mut penguin_rx) = penguin.space::<SecretData>(topic).await?;
@@ -374,17 +384,18 @@ mod spaces_repair_task {
     use p2panda_spaces::SpaceEvent;
     use tokio_stream::StreamExt;
 
-    use super::SecretData;
+    use super::{SecretData, spawn_node};
 
     #[tokio::test]
     async fn sync_repair_space() {
         setup_logging();
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
         let mut panda_system_rx = panda.event_stream().await.unwrap();
-        let penguin = p2panda::spawn().await.unwrap();
+        let penguin = spawn_node(network_id).await;
         let mut penguin_system_rx = penguin.event_stream().await.unwrap();
 
         // Penguin creates a group before subscribing to the space.
@@ -461,11 +472,12 @@ mod spaces_repair_task {
     async fn live_repair_space() {
         setup_logging();
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
         let mut panda_system_rx = panda.event_stream().await.unwrap();
-        let penguin = p2panda::spawn().await.unwrap();
+        let penguin = spawn_node(network_id).await;
 
         // Penguin subscribes to the space.
         let (_penguin_space, mut penguin_rx) = penguin.space::<SecretData>(topic).await.unwrap();
@@ -540,15 +552,16 @@ mod spaces_api_validation {
     use p2panda_spaces::SpaceEvent;
     use tokio_stream::StreamExt;
 
-    use super::SecretData;
+    use super::{SecretData, spawn_node};
 
     #[tokio::test]
     async fn api_validation() {
         setup_logging();
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
 
         let (panda_space, mut panda_rx) = panda.create_space::<String>(topic).await.unwrap();
 
@@ -575,7 +588,7 @@ mod spaces_api_validation {
         );
 
         // Tiger subscribes to the space.
-        let tiger = p2panda::spawn().await.unwrap();
+        let tiger = spawn_node(network_id).await;
         let (tiger_space, mut tiger_rx) = tiger.space::<String>(topic).await.unwrap();
 
         while let Some(event) = panda_rx.next().await {
@@ -631,9 +644,11 @@ mod spaces_api_validation {
     async fn groups_api_validation() {
         setup_logging();
 
-        let panda = p2panda::spawn().await.unwrap();
-        let lion = p2panda::spawn().await.unwrap();
-        let tiger = p2panda::spawn().await.unwrap();
+        let network_id = Topic::random().into();
+
+        let panda = spawn_node(network_id).await;
+        let lion = spawn_node(network_id).await;
+        let tiger = spawn_node(network_id).await;
 
         let topic = Topic::random();
 
@@ -730,31 +745,32 @@ mod spaces_api_validation {
 }
 
 mod spaces_events {
+    use p2panda::Topic;
     use p2panda::spaces::{GroupEvent, InnerGroupEvent};
     use p2panda::streams::SystemEvent;
     use p2panda_auth::AccessLevel;
     use p2panda_core::test_utils::setup_logging;
     use tokio_stream::StreamExt;
 
-    use super::SecretData;
+    use super::{SecretData, spawn_node};
 
     #[tokio::test]
     async fn group_events() {
         setup_logging();
 
-        use p2panda::Topic;
+        let network_id = Topic::random().into();
 
-        let panda = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
         let mut panda_system_rx = panda.event_stream().await.unwrap();
 
         let topic = Topic::random();
 
         // Create a space with only us inside. Having a space in this test is only required to sync
         // the group operations.
-        let (_panda_space, _panda_rx) = panda.create_space::<SecretData>(topic).await.unwrap();
+        let (panda_space, _panda_rx) = panda.create_space::<SecretData>(topic).await.unwrap();
 
-        let penguin_laptop = p2panda::spawn().await.unwrap();
-        let penguin_mobile = p2panda::spawn().await.unwrap();
+        let penguin_laptop = spawn_node(network_id).await;
+        let penguin_mobile = spawn_node(network_id).await;
         let mut penguin_laptop_system_rx = penguin_laptop.event_stream().await.unwrap();
 
         let (_penguin_laptop_space, _penguin_laptop_rx) =
@@ -808,11 +824,13 @@ mod spaces_events {
 
         // Penguin receives the team group event on their system stream.
         loop {
+            let event = penguin_laptop_system_rx.next().await;
+
             if let Some(SystemEvent::Groups {
                 group_id,
                 inner: InnerGroupEvent::Created { .. },
                 ..
-            }) = penguin_laptop_system_rx.next().await
+            }) = event
             {
                 if group_id == team_group.id() {
                     break;
@@ -902,7 +920,7 @@ mod filtered_messages {
     use p2panda_core::test_utils::setup_logging;
     use tokio_stream::StreamExt;
 
-    use super::SecretData;
+    use super::{SecretData, spawn_node};
 
     #[tokio::test]
     async fn concurrently_removed_members_filtered() {
@@ -911,10 +929,11 @@ mod filtered_messages {
         use p2panda::Topic;
         use p2panda_auth::AccessLevel;
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await.unwrap();
-        let penguin = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
+        let penguin = spawn_node(network_id).await;
 
         // Panda creates a space.
         let (panda_space, mut panda_rx) = panda.create_space::<SecretData>(topic).await.unwrap();
@@ -1031,11 +1050,12 @@ mod filtered_messages {
         use p2panda::Topic;
         use p2panda_auth::AccessLevel;
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let panda = p2panda::spawn().await.unwrap();
-        let penguin = p2panda::spawn().await.unwrap();
-        let tiger = p2panda::spawn().await.unwrap();
+        let panda = spawn_node(network_id).await;
+        let penguin = spawn_node(network_id).await;
+        let tiger = spawn_node(network_id).await;
 
         // Panda creates a space.
         let (panda_space, mut panda_rx) = panda.create_space::<SecretData>(topic).await.unwrap();
@@ -1131,6 +1151,8 @@ mod members {
     use p2panda_core::test_utils::setup_logging;
     use tokio_stream::StreamExt;
 
+    use crate::spawn_node;
+
     // 1. Node A creates space S with {A, B, C, D} inside
     // 2. Node B removes C from S
     //
@@ -1141,12 +1163,13 @@ mod members {
     async fn indirect_members_log_sync() {
         setup_logging();
 
+        let network_id = Topic::random().into();
         let topic = Topic::random();
 
-        let node_a = p2panda::spawn().await.unwrap();
-        let node_b = p2panda::spawn().await.unwrap();
-        let node_c = p2panda::spawn().await.unwrap();
-        let node_d = p2panda::spawn().await.unwrap();
+        let node_a = spawn_node(network_id).await;
+        let node_b = spawn_node(network_id).await;
+        let node_c = spawn_node(network_id).await;
+        let node_d = spawn_node(network_id).await;
 
         let (node_a_space, mut node_a_rx) = node_a.create_space::<String>(topic).await.unwrap();