+++ CHANGELOG.md
@@ -46,6 +46,8 @@ Highlights are marked with a pancake 🥞
 - net: Update ractor to 0.16.2 [#1324](https://github.com/p2panda/p2panda/pull/1324)
 - sync: Tracing for sync state machine using instrument span [#1289](https://github.com/p2panda/p2panda/pull/1289)
 - node: Move repair task out of stream [#1347](https://github.com/p2panda/p2panda/pull/1347)
+- node: Increase stream buffer size [#1360](https://github.com/p2panda/p2panda/pull/1360)
+- node: Use random network ids in tests [#1360](https://github.com/p2panda/p2panda/pull/1360)
 
 ### Fixed